requirements.txt
torch==1.9.0+cu102
numpy==1.18.5
six==1.16.0
~pencv_contrib_python==4.5.3.56
tensorflow==2.3.0
lmdb==1.2.1
matplotlib==3.3.4
scikit_image==0.17.2
config==0.5.0.post0
torchvision==0.10.0+cu102
scipy==1.4.1
natsort==7.1.1
Pillow==8.3.1
skimage==0.0
